tests/wpt/web-platform-tests/bluetooth/idl/idl-NavigatorBluetooth.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';
const test_desc = '[SameObject] test for navigator.bluetooth';

test(() => {
  assert_true('bluetooth' in navigator, 'navigator.bluetooth exists.');
}, 'navigator.bluetooth IDL test');

test(() => {
  assert_equals(navigator.bluetooth, navigator.bluetooth);
}, test_desc);
</script>
